Raise exception is population size is invalid
Creating a population with a `population_size` of an odd number results
in `Population#make_next_generation!` going into an infinite loop, as
the membership of the new population increases in size by pairs, passing
the desired population size and continuing indefinitely.

While this issue made sense after considering what breeding pairs does,
it did result in several futile attempts to track down what behavior was
causing an infinite loop.

This change adds two checks, the first that the desired population size
is even, and the second that the number is positive. If either of those
checks fails, we halt execution and raise a runtime exception.

Additionally deleted a line of code that assigned a temporary variable
that was never accessed again.