Don't create a new string for frequent breeding comparison
Each time we create a new member a check is made to check if the
organism's class is `Darwinning::Organism`. This check can be performed
by comparing the class names with existing constants rather than
creating a new string for each comparison, which uses more memory and
may be less performant across hundreds of thousands of operations during
evolution.