Update Population#make_next_generation! to handle odd population sizes
Context: There was an infinite loop prior to validation check added in:
https://github.com/dorkrawk/darwinning/commit/95c6c4c9bf853b0cb0ebdf939ae912d174b66f13

This commit allows an extra pair to added to the population, but terminates
the loop if there are more new members than desired. There is a follow up
check to `pop` off the last added member if that occurred.

* Kept the check in place that the population size needs to a positive
number since that seems sensible.