Fix bug within the Mutation evolution type
The existing code did not update the member's genotypes when
re-expressing a gene.

This change borrows the assignment techniques from the `Reproduction`
evolution type which observes the distinction between the two styles:
... subclass Organism and the 'include Darwinning' style.

In the case of the 'include' style
... the member's #genotypes method returns a new Hash each time its called
... see: https://github.com/dorkrawk/darwinning/blob/ef958c6b22a744a0259606901442ffb7c785cf84/lib/darwinning.rb#L57
And so we need to make use of the contract that there are accessors
for each of the genes, and send messages to the setter methods.

In the case of the Organism subclass
... we can modify the member.genotypes directly,
... but we need to use the gene object as the key to update the value.

This commit also adds two supporting files to spec/classes
... they each have only one gene to facilitate testing.